+++ models/topics.go
@@ -3,6 +3,7 @@ package models
 import (
 	"time"
 
+	"github.com/lib/pq"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -23,18 +24,30 @@ type NewTopicRequestData struct {
 	Body string `json:"body" bson:"body"`
 }
 
-// 此处不使用 ID 是为了避免插入时覆盖 mongodb 自增ID
+// 帖子表
 type Topic struct {
-	// 作者
-	Author User `json:"author" bson:"author"`
+	// 帖子 ID
+	ID int64 `gorm:"primaryKey;autoIncrement"`
+	// 作者 ID（User 外键：即不是 Topic 自身的属性，为了关联 User 而存在。）
+	AuthorID int64 `gorm:"not null"`
+	// 作者实例（GORM 预加载用，如果外键引用的 User 中的属性不是 `ID`，需要使用 `references` 手动指定。）
+	Author User `gorm:"foreignKey:AuthorID;"`
 	// 版块（版块ID）
-	Plate string `json:"plate" bson:"plate"`
+	PlateID string `gorm:"not null"`
+	// 版块实例
+	Plate Plate `gorm:"foreignKey:PlateID"`
 	// 标题
-	Title string `json:"title" bson:"title"`
+	Title string `gorm:"not null"`
+	// 帖子类型
+	TopicType string `gorm:"size:20;not null"`
+	// 预览图链接
+	ThumbnailLink string `gorm:"not null"`
 	// TAG
-	Tag []string `json:"tag" bson:"tag"`
-	// 创建日
-	CreationDate time.Time `json:"creationDate" bson:"creationDate"`
+	Tag pq.StringArray `gorm:"type:varchar(20)[]"`
+	// 创建时间
+	CreatedAt time.Time `gorm:"not null"`
+	// 更新时间
+	UpdatedAt time.Time `gorm:"not null"`
 }
 
 type TopicWidthID struct {