+++ repositories/topics.go
@@ -1,11 +1,60 @@
 package repositories
 
 import (
+	"github.com/NSDN/nya-server/context"
 	"github.com/NSDN/nya-server/models"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
+	"gorm.io/gorm"
 )
+
+type TopicRepository struct {
+	context *context.AppContext
+}
+
+func NewTopicRepository(context *context.AppContext) *TopicRepository {
+	return &TopicRepository{context: context}
+}
+
+// 获取帖子列表 - 数据库
+func (repository *TopicRepository) GetTopics(plateID string) (
+	[]models.Topic,
+	error,
+) {
+	var topics []models.Topic
+
+	// 传给 `Preload` 的字符串是 `models.Topic` 中预留填充关联结果的属性名。
+	result := repository.context.DB.
+		Where("plate_id = ?", plateID).
+		Select("id", "title", "thumbnail_link", "updated_at").
+		Order("updated_at DESC").
+		Find(&topics)
+
+	if result.Error != nil {
+		return nil, result.Error
+	}
+
+	return topics, nil
+}
+
+// 获取帖子列表 - 数据库
+func GetTopics(db *gorm.DB, plateID string) ([]models.Topic, error) {
+	var topics []models.Topic
+
+	// 传给 `Preload` 的字符串是 `models.Topic` 中预留填充关联结果的属性名。
+	result := db.
+		Where("plate_id <> ?", plateID).
+		Select("id", "title", "thumbnail_link", "updated_at").
+		Order("updated_at DESC").
+		Find(&topics)
+
+	if result.Error != nil {
+		return nil, result.Error
+	}
+
+	return topics, nil
+}
 
 // 获取帖子集合
 func getTopicCollection() *mongo.Collection {
@@ -26,7 +75,7 @@ func CreateTopic(topic *models.Topic) (*mongo.InsertOneResult, error) {
 }
 
 // 获取帖子列表 - 数据库
-func GetTopics(plate string) (*[]models.TopicWidthID, error) {
+func GetTopicsOld(plate string) (*[]models.TopicWidthID, error) {
 	collection := getTopicCollection()
 
 	topics, err := findDataFromCollection(