+++ services/plates.go
@@ -4,15 +4,27 @@ import (
 	"log"
 
 	"github.com/NSDN/nya-server/constants"
+	"github.com/NSDN/nya-server/context"
 	"github.com/NSDN/nya-server/models"
 	"github.com/NSDN/nya-server/repositories"
-	"gorm.io/gorm"
 )
 
+type PlateService struct {
+	context    *context.AppContext
+	repository *repositories.PlateRepository
+}
+
+func NewPlateService(context *context.AppContext) *PlateService {
+	return &PlateService{
+		context:    context,
+		repository: repositories.NewPlateRepository(context),
+	}
+}
+
 // 创建版块列表 - 服务
-func InitPlateList(db *gorm.DB) {
+func (service *PlateService) InitPlateList() {
 	// 从数据库中获取版块列表
-	plates, err := GetPlateList(db)
+	plates, err := service.GetPlates()
 
 	if err != nil {
 		log.Fatal(err)
@@ -26,26 +38,32 @@ func InitPlateList(db *gorm.DB) {
 	// 如果没有既存列表，则插入列表
 	newPlates := []models.Plate{
 		{
-			ID:         "localization",
-			Name:       "喵玉汉化馆",
-			Background: "https://i.imgur.com/ohQuzivl.jpg",
-			PageType:   constants.COMIC,
+			ID:          "localization",
+			Name:        "喵玉汉化馆",
+			Description: "",
+			Background:  "https://i.imgur.com/ohQuzivl.jpg",
+			PageType:    constants.COMIC,
+			SortOrder:   0,
 		},
 		{
-			ID:         "music",
-			Name:       "喵玉咏唱组",
-			Background: "https://i.imgur.com/IHo7tTyl.jpg",
-			PageType:   constants.ARTICLE,
+			ID:          "music",
+			Name:        "喵玉咏唱组",
+			Description: "",
+			Background:  "https://i.imgur.com/IHo7tTyl.jpg",
+			PageType:    constants.ARTICLE,
+			SortOrder:   1,
 		},
 		{
-			ID:         "chat",
-			Name:       "魔女的茶会",
-			Background: "https://i.imgur.com/JsWkJ4jl.jpg",
-			PageType:   constants.ARTICLE,
+			ID:          "chat",
+			Name:        "魔女的茶会",
+			Description: "",
+			Background:  "https://i.imgur.com/JsWkJ4jl.jpg",
+			PageType:    constants.ARTICLE,
+			SortOrder:   2,
 		},
 	}
 
-	_, err = repositories.InitPlateList(db, newPlates)
+	_, err = service.repository.InitPlateList(newPlates)
 
 	if err != nil {
 		log.Fatal(err)
@@ -53,6 +71,6 @@ func InitPlateList(db *gorm.DB) {
 }
 
 // 获取版块列表 - 服务
-func GetPlateList(db *gorm.DB) ([]models.Plate, error) {
-	return repositories.GetPlateList(db)
+func (service *PlateService) GetPlates() ([]models.Plate, error) {
+	return service.repository.GetPlates()
 }