+++ services/topics.go
@@ -1,11 +1,50 @@
 package services
 
 import (
+	"strconv"
+
+	"github.com/NSDN/nya-server/context"
+	"github.com/NSDN/nya-server/dto"
 	"github.com/NSDN/nya-server/models"
 	"github.com/NSDN/nya-server/repositories"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 )
+
+type TopicService struct {
+	context    *context.AppContext
+	repository *repositories.TopicRepository
+}
+
+func NewTopicService(context *context.AppContext) *TopicService {
+	return &TopicService{
+		context:    context,
+		repository: repositories.NewTopicRepository(context),
+	}
+}
+
+// 获取帖子列表 - 服务
+func (service *TopicService) GetTopics(plateID string) ([]dto.TopicListItemDTO, error) {
+	topics, err := service.repository.GetTopics(plateID)
+
+	if err != nil {
+		return nil, err
+	}
+
+	result := make([]dto.TopicListItemDTO, 0, len(topics))
+
+	for _, topic := range topics {
+		// TODO: 将 `strconv.FormatInt` 更改为真的编码方法。
+		result = append(result, dto.TopicListItemDTO{
+			ID:            strconv.FormatInt(topic.ID, 10),
+			Title:         topic.Title,
+			ThumbnailLink: topic.ThumbnailLink,
+			UpdatedAt:     topic.UpdatedAt,
+		})
+	}
+
+	return result, nil
+}
 
 // 创建帖子 - 服务
 //
@@ -13,11 +52,10 @@ import (
 // 楼层使用创建帖子信息时由数据库自增生成的 ID 作为唯一键。
 func CreateTopic(request *models.NewTopicRequestData) (bool, error) {
 	topic := models.Topic{
-		Author:       request.Author,
-		Plate:        request.Plate,
-		Title:        request.Title,
-		Tag:          request.Tag,
-		CreationDate: request.CreationDate,
+		Author:  request.Author,
+		PlateID: request.Plate,
+		Title:   request.Title,
+		Tag:     request.Tag,
 	}
 
 	insertResult, err := repositories.CreateTopic(&topic)
@@ -59,8 +97,8 @@ func CreateTopicFloors(
 // 获取帖子列表 - 服务
 //
 // 从数据库中找出所有帖子，然后根据传入的版块路由名来过筛选出当前版块的帖子列表。
-func GetTopics(plate string) (*[]models.TopicWidthID, error) {
-	topics, err := repositories.GetTopics(plate)
+func GetTopicsOld(plate string) (*[]models.TopicWidthID, error) {
+	topics, err := repositories.GetTopicsOld(plate)
 
 	if err != nil {
 		return nil, err