+++ CHANGELOG.md
@@ -6,11 +6,14 @@
 
 - 创建更新日志。
 <!-- - 创建贡献指南。 -->
+- 为版块表追加 `sort_order` 列。
 
 ### Change
 
 - 将数据库从 MongoDB 换成 PostgreSQL。
 - 重构版块整体 API。
+- 重构帖子表定义从 MongoDB 至 PostgreSQL。
+- 重构获取帖子列表 API。
 
 [Unreleased]: https://github.com/NSDN/nya-client/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/NSDN/nya-client/release/tag/v1.0.0